random updates on file structures
